+++ include/random.h
@@ -1,2 +1,2 @@
-long randlong();
-long randlongrange(long lower, long upper);
+__int128_t randlong();
+__int128_t randlongrange(__int128_t lower, __int128_t upper);
