+++ include/rsa.h
@@ -1,18 +1,18 @@
 struct PublicKey {
-    unsigned long e;
-    unsigned long N;
+    __uint128_t e;
+    __uint128_t N;
 };
 
 struct PrivateKey {
     struct PublicKey pub;
-    unsigned long p;
-    unsigned long q;
-    unsigned long d;
+    __uint128_t p;
+    __uint128_t q;
+    __uint128_t d;
 };
 
-long encryptc(struct PublicKey key, char m);
-char decryptc(struct PrivateKey key, long c);
-long* encrypt(struct PublicKey key, char* m);
-char* decrypt(struct PrivateKey key, long* c);
+__int128_t encryptc(struct PublicKey key, char m);
+char decryptc(struct PrivateKey key, __int128_t c);
+__int128_t* encrypt(struct PublicKey key, char* m);
+char* decrypt(struct PrivateKey key, __int128_t* c);
 char* pubkey_to_ascii(struct PublicKey key);
 char* privkey_to_ascii(struct PrivateKey key);
