+++ rsa.c
@@ -6,23 +6,23 @@
 #include <string.h>
 #include <stdio.h>
 
-long encryptc(struct PublicKey key, char m) {
+__int128_t encryptc(struct PublicKey key, char m) {
     return modpower(m,key.e, key.N);
 }
 
-char decryptc(struct PrivateKey key, long c) {
+char decryptc(struct PrivateKey key, __int128_t c) {
     return modpower(c,key.d,key.pub.N);
 }
 
-long* encrypt(struct PublicKey key, char* m) {
+__int128_t* encrypt(struct PublicKey key, char* m) {
     int length = strlen(m);
-    long* c = malloc(length * sizeof(long) +1);
+    __int128_t* c = malloc(length * sizeof(__int128_t) +1);
     for (int i=0; i < length; i++) c[i] = encryptc(key, m[i]);
     c[length] = 0;
     return c;
 }
 
-char* decrypt(struct PrivateKey key, long* c) {
+char* decrypt(struct PrivateKey key, __int128_t* c) {
     int length = 0;
     for (;c[length] != 0; length++);
     char *m = malloc(length);
